Refs 11636 adding poc removing logic

+++ app/models/person.rb
@@ -699,7 +699,7 @@ class Person
       rescue
         return false, 'Person not found'
       end
-      if role = person.employer_staff_roles.detect{|role| role.employer_profile_id.to_s == employer_profile_id.to_s}
+      if role = person.employer_staff_roles.detect{|role| role.employer_profile_id.to_s == employer_profile_id.to_s && !role.is_closed?}
         role.update_attributes!(:aasm_state => :is_closed)
        role.destroy!
         return true, 'Employee Staff Role is inactive'
